+++ CHANGELOG.md
@@ -10,7 +10,8 @@ None
 - Use Bract 0.6.1 components
 - Use Aleph 0.4.6
 - Use `:repl-options` to set port to 3001
-- [Todo] Use reloadable var for homepage handler
+- Use reloadable homepage handler
+- Fix `/items/:id` URI handler call
 
 
 ## 0.1.1 / 2018-May-24